projects/inoahglobal/requirements.txt
# iNoah Global - Shared Infrastructure
# These are the core dependencies needed by the shared package

requests>=2.31.0
Pillow>=10.0.0
chromadb>=0.4.0
